refactor(hir): split widget_decl.rs under the 2000-line file-size cap
widget_decl.rs reached 2372 lines — #6513's reloadPolicy work pushed it
over the CI `check_file_size.sh` cap, reddening the lint job's file-size
step on every PR (previously masked behind the benchmark-freshness step).
Pure mechanical split, no logic change:

- widget_decl/reactive_animate.rs — collect_state_value_reads +
  try_desugar_reactive_animate (reactive-animation desugaring).
  try_desugar_reactive_animate is re-exported from widget_decl.rs as
  pub(crate) so lower/mod.rs's `pub(crate) use widget_decl::*` still
  resolves it for expr_call/prescans.rs.
- widget_decl/tests.rs — the #[cfg(test)] module.

widget_decl.rs drops to 1665 lines. perry-hir compiles, fmt --check is
clean, the 18 moved tests pass, and check_file_size.sh is green.
