+++ crates/perry-hir/src/lower/widget_decl.rs
@@ -14,271 +14,8 @@ use swc_ecma_ast as ast;
 use super::*;
 use crate::ir::*;
 
-/// Walk an AST expression and collect identifiers used as `<ident>.value`
-/// where `<ident>` resolves to a `perry/ui` State native instance. Callers
-/// use the collected names to register `stateOnChange` subscribers.
-///
-/// Covers the expression shapes most commonly found in animation arguments:
-/// ternaries, binary/logical ops, parens, template literals, unary,
-/// assignment RHS, call args, array/object literals, and member reads. The
-/// catch-all silently skips unhandled shapes — worst case, a state read
-/// inside an exotic expression just won't trigger reactivity (same
-/// conservative failure mode as #104's template walker).
-fn collect_state_value_reads(ctx: &LoweringContext, expr: &ast::Expr, out: &mut Vec<String>) {
-    match expr {
-        ast::Expr::Member(member) => {
-            // `<ident>.value` where ident is a registered State.
-            if let ast::MemberProp::Ident(prop) = &member.prop {
-                if prop.sym.as_ref() == "value" {
-                    if let ast::Expr::Ident(obj) = member.obj.as_ref() {
-                        let name = obj.sym.to_string();
-                        if matches!(
-                            ctx.lookup_native_instance(&name),
-                            Some(("perry/ui", "State"))
-                        ) && !out.contains(&name)
-                        {
-                            out.push(name);
-                            return;
-                        }
-                    }
-                }
-            }
-            collect_state_value_reads(ctx, member.obj.as_ref(), out);
-        }
-        ast::Expr::Paren(p) => collect_state_value_reads(ctx, &p.expr, out),
-        ast::Expr::Cond(c) => {
-            collect_state_value_reads(ctx, &c.test, out);
-            collect_state_value_reads(ctx, &c.cons, out);
-            collect_state_value_reads(ctx, &c.alt, out);
-        }
-        ast::Expr::Bin(b) => {
-            collect_state_value_reads(ctx, &b.left, out);
-            collect_state_value_reads(ctx, &b.right, out);
-        }
-        ast::Expr::Unary(u) => collect_state_value_reads(ctx, &u.arg, out),
-        ast::Expr::Tpl(t) => {
-            for e in &t.exprs {
-                collect_state_value_reads(ctx, e, out);
-            }
-        }
-        ast::Expr::Call(c) => {
-            if let ast::Callee::Expr(ce) = &c.callee {
-                collect_state_value_reads(ctx, ce, out);
-            }
-            for a in &c.args {
-                collect_state_value_reads(ctx, &a.expr, out);
-            }
-        }
-        ast::Expr::Array(a) => {
-            for el in a.elems.iter().flatten() {
-                collect_state_value_reads(ctx, &el.expr, out);
-            }
-        }
-        ast::Expr::Seq(s) => {
-            for e in &s.exprs {
-                collect_state_value_reads(ctx, e, out);
-            }
-        }
-        ast::Expr::TsNonNull(n) => collect_state_value_reads(ctx, &n.expr, out),
-        ast::Expr::TsAs(a) => collect_state_value_reads(ctx, &a.expr, out),
-        ast::Expr::TsTypeAssertion(a) => collect_state_value_reads(ctx, &a.expr, out),
-        _ => {}
-    }
-}
-
-/// Desugar `widget.animateOpacity(<expr>, dur)` / `.animatePosition(...)`
-/// into an IIFE that runs the initial animation and registers a
-/// `stateOnChange` subscriber per `State` read in the args, so toggling the
-/// state re-fires the animation.
-///
-/// Generated HIR shape (animateOpacity with one state dependency):
-/// ```text
-/// (() => {
-///     const __h = <widget>;
-///     widgetAnimateOpacity(__h, target, dur);       // initial
-///     stateOnChange(state1, (__v) => widgetAnimateOpacity(__h, fresh_target, dur));
-///     return undefined;
-/// })()
-/// ```
-///
-/// Like the reactive-Text desugar (#104), the target expression is re-lowered
-/// for the subscriber body so it reads the *current* state value at fire time.
-pub(super) fn try_desugar_reactive_animate(
-    ctx: &mut LoweringContext,
-    call: &ast::CallExpr,
-) -> Result<Option<Expr>> {
-    let ast::Callee::Expr(callee_expr) = &call.callee else {
-        return Ok(None);
-    };
-    let ast::Expr::Member(member) = callee_expr.as_ref() else {
-        return Ok(None);
-    };
-    let ast::MemberProp::Ident(prop) = &member.prop else {
-        return Ok(None);
-    };
-    let (method_name, expected_arity) = match prop.sym.as_ref() {
-        "animateOpacity" => ("widgetAnimateOpacity", 2),
-        "animatePosition" => ("widgetAnimatePosition", 3),
-        _ => return Ok(None),
-    };
-    if call.args.iter().any(|a| a.spread.is_some()) {
-        return Ok(None);
-    }
-    if call.args.len() != expected_arity {
-        return Ok(None);
-    }
-
-    // Collect unique state names whose `.value` is read anywhere in the args.
-    // Preserving insertion order keeps subscriber registration deterministic.
-    let mut state_names: Vec<String> = Vec::new();
-    for arg in &call.args {
-        collect_state_value_reads(ctx, &arg.expr, &mut state_names);
-    }
-    if state_names.is_empty() {
-        return Ok(None);
-    }
-
-    // Lower the receiver once; store in an IIFE local so the initial call and
-    // every subscriber share the same widget handle without re-evaluating
-    // side-effectful receiver expressions.
-    let widget_expr = lower_expr(ctx, member.obj.as_ref())?;
-
-    let outer_func_id = ctx.fresh_func();
-    let outer_scope = ctx.enter_scope();
-    let widget_id = ctx.define_local("__perry_anim_widget".to_string(), Type::Any);
-
-    let mut outer_body: Vec<Stmt> = Vec::new();
-    outer_body.push(Stmt::Let {
-        id: widget_id,
-        name: "__perry_anim_widget".to_string(),
-        ty: Type::Any,
-        mutable: false,
-        init: Some(widget_expr),
-    });
-
-    let mut initial_args: Vec<Expr> = Vec::with_capacity(expected_arity + 1);
-    initial_args.push(Expr::LocalGet(widget_id));
-    for a in &call.args {
-        initial_args.push(lower_expr(ctx, &a.expr)?);
-    }
-    outer_body.push(Stmt::Expr(Expr::NativeMethodCall {
-        module: "perry/ui".to_string(),
-        method: method_name.to_string(),
-        object: None,
-        args: initial_args,
-        class_name: None,
-    }));
-
-    for state_name in &state_names {
-        let state_local = ctx
-            .lookup_local(state_name)
-            .ok_or_else(|| anyhow!("reactive animate: state '{}' not in scope", state_name))?;
-
-        let inner_func_id = ctx.fresh_func();
-        let inner_scope = ctx.enter_scope();
-        let v_param_id = ctx.define_local("__v".to_string(), Type::Any);
-        let v_param = Param {
-            id: v_param_id,
-            name: "__v".to_string(),
-            ty: Type::Any,
-            default: None,
-            decorators: Vec::new(),
-            is_rest: false,
-            arguments_object: None,
-        };
-
-        let mut fresh_args: Vec<Expr> = Vec::with_capacity(expected_arity + 1);
-        fresh_args.push(Expr::LocalGet(widget_id));
-        for a in &call.args {
-            fresh_args.push(lower_expr(ctx, &a.expr)?);
-        }
-        let animate_call = Expr::NativeMethodCall {
-            module: "perry/ui".to_string(),
-            method: method_name.to_string(),
-            object: None,
-            args: fresh_args,
-            class_name: None,
-        };
-        let inner_body = vec![Stmt::Expr(animate_call)];
-        ctx.exit_scope(inner_scope);
-
-        let mut inner_refs = Vec::new();
-        let mut inner_visited = std::collections::HashSet::new();
-        for stmt in &inner_body {
-            collect_local_refs_stmt(stmt, &mut inner_refs, &mut inner_visited);
-        }
-        let mut inner_captures: Vec<LocalId> = inner_refs
-            .into_iter()
-            .filter(|id| *id != v_param_id)
-            .collect();
-        inner_captures.sort();
-        inner_captures.dedup();
-        inner_captures = ctx.filter_module_level_captures(inner_captures);
-
-        let inner_closure = Expr::Closure {
-            func_id: inner_func_id,
-            params: vec![v_param],
-            return_type: Type::Any,
-            body: inner_body,
-            captures: inner_captures,
-            mutable_captures: Vec::new(),
-            captures_this: false,
-            captures_new_target: false,
-            enclosing_class: None,
-            is_arrow: false,
-            is_async: false,
-            is_generator: false,
-            is_strict: ctx.current_strict,
-        };
-
-        outer_body.push(Stmt::Expr(Expr::NativeMethodCall {
-            module: "perry/ui".to_string(),
-            method: "stateOnChange".to_string(),
-            object: None,
-            args: vec![Expr::LocalGet(state_local), inner_closure],
-            class_name: None,
-        }));
-    }
-
-    outer_body.push(Stmt::Return(Some(Expr::Undefined)));
-    ctx.exit_scope(outer_scope);
-
-    let mut outer_refs = Vec::new();
-    let mut outer_visited = std::collections::HashSet::new();
-    for stmt in &outer_body {
-        collect_local_refs_stmt(stmt, &mut outer_refs, &mut outer_visited);
-    }
-    let mut outer_captures: Vec<LocalId> = outer_refs
-        .into_iter()
-        .filter(|id| *id != widget_id)
-        .collect();
-    outer_captures.sort();
-    outer_captures.dedup();
-    outer_captures = ctx.filter_module_level_captures(outer_captures);
-
-    let outer_closure = Expr::Closure {
-        func_id: outer_func_id,
-        params: vec![],
-        return_type: Type::Any,
-        body: outer_body,
-        captures: outer_captures,
-        mutable_captures: Vec::new(),
-        captures_this: false,
-        captures_new_target: false,
-        enclosing_class: None,
-        is_arrow: false,
-        is_async: false,
-        is_generator: false,
-        is_strict: ctx.current_strict,
-    };
-
-    Ok(Some(Expr::Call {
-        callee: Box::new(outer_closure),
-        args: vec![],
-        type_args: vec![],
-        byte_offset: 0,
-    }))
-}
+mod reactive_animate;
+pub(crate) use reactive_animate::try_desugar_reactive_animate;
 
 /// Try to lower a Widget({...}) call from perry/widget into a WidgetDecl.
 /// Returns Some(WidgetDecl) if this is a widget declaration, None otherwise.
@@ -1925,448 +1662,4 @@ fn parse_placeholder_value(expr: &ast::Expr) -> WidgetPlaceholderValue {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use swc_common::DUMMY_SP;
-    use swc_ecma_ast as ast;
-
-    fn str_lit(value: &str) -> ast::Expr {
-        ast::Expr::Lit(ast::Lit::Str(ast::Str {
-            span: DUMMY_SP,
-            value: value.into(),
-            raw: None,
-        }))
-    }
-
-    fn key_value(key: &str, value: ast::Expr) -> ast::PropOrSpread {
-        ast::PropOrSpread::Prop(Box::new(ast::Prop::KeyValue(ast::KeyValueProp {
-            key: ast::PropName::Ident(ast::IdentName {
-                span: DUMMY_SP,
-                sym: key.into(),
-            }),
-            value: Box::new(value),
-        })))
-    }
-
-    fn array_lit(elems: Vec<ast::Expr>) -> ast::Expr {
-        ast::Expr::Array(ast::ArrayLit {
-            span: DUMMY_SP,
-            elems: elems
-                .into_iter()
-                .map(|e| {
-                    Some(ast::ExprOrSpread {
-                        spread: None,
-                        expr: Box::new(e),
-                    })
-                })
-                .collect(),
-        })
-    }
-
-    fn object_lit(props: Vec<ast::PropOrSpread>) -> ast::Expr {
-        ast::Expr::Object(ast::ObjectLit {
-            span: DUMMY_SP,
-            props,
-        })
-    }
-
-    #[test]
-    fn primitive_spec_recognized_scalars() {
-        assert!(matches!(
-            parse_entry_field_primitive_spec("number"),
-            WidgetFieldType::Number
-        ));
-        assert!(matches!(
-            parse_entry_field_primitive_spec("boolean"),
-            WidgetFieldType::Boolean
-        ));
-        assert!(matches!(
-            parse_entry_field_primitive_spec("string"),
-            WidgetFieldType::String
-        ));
-    }
-
-    #[test]
-    fn primitive_spec_unknown_falls_back_to_string() {
-        assert!(matches!(
-            parse_entry_field_primitive_spec("Date"),
-            WidgetFieldType::String
-        ));
-        assert!(matches!(
-            parse_entry_field_primitive_spec(""),
-            WidgetFieldType::String
-        ));
-    }
-
-    #[test]
-    fn primitive_spec_optional_and_array_suffixes() {
-        assert!(matches!(
-            parse_entry_field_primitive_spec("number?"),
-            WidgetFieldType::Optional(inner) if matches!(*inner, WidgetFieldType::Number)
-        ));
-        assert!(matches!(
-            parse_entry_field_primitive_spec("string[]"),
-            WidgetFieldType::Array(inner) if matches!(*inner, WidgetFieldType::String)
-        ));
-        // `'boolean[]?'` — array suffix is parsed first, then optional.
-        let nested = parse_entry_field_primitive_spec("boolean[]?");
-        let WidgetFieldType::Optional(o) = nested else {
-            panic!("expected Optional, got {:?}", nested);
-        };
-        let WidgetFieldType::Array(a) = *o else {
-            panic!("expected Optional<Array>, got Optional<other>");
-        };
-        assert!(matches!(*a, WidgetFieldType::Boolean));
-    }
-
-    #[test]
-    fn entry_field_value_spec_object_literal_becomes_object() {
-        // Inline `{ url: 'string', clicks: 'number' }`.
-        let expr = object_lit(vec![
-            key_value("url", str_lit("string")),
-            key_value("clicks", str_lit("number")),
-        ]);
-        let ty = parse_entry_field_value_spec(&expr);
-        let WidgetFieldType::Object(fields) = ty else {
-            panic!("expected Object, got {:?}", ty);
-        };
-        assert_eq!(fields.len(), 2);
-        assert_eq!(fields[0].0, "url");
-        assert!(matches!(fields[0].1, WidgetFieldType::String));
-        assert_eq!(fields[1].0, "clicks");
-        assert!(matches!(fields[1].1, WidgetFieldType::Number));
-    }
-
-    #[test]
-    fn entry_field_value_spec_array_of_objects_does_not_collapse_to_string() {
-        // Inline `[{ url: 'string', clicks: 'number' }]`. This is the
-        // exact shape that used to collapse to `String` in the old
-        // parser, breaking SwiftUI `ForEach` over the field.
-        let inner = object_lit(vec![
-            key_value("url", str_lit("string")),
-            key_value("clicks", str_lit("number")),
-        ]);
-        let expr = array_lit(vec![inner]);
-        let ty = parse_entry_field_value_spec(&expr);
-        let WidgetFieldType::Array(elem) = ty else {
-            panic!("expected Array, got {:?}", ty);
-        };
-        let WidgetFieldType::Object(fields) = *elem else {
-            panic!("expected Array<Object>, got Array<other>");
-        };
-        assert_eq!(fields.len(), 2);
-        assert_eq!(fields[0].0, "url");
-        assert_eq!(fields[1].0, "clicks");
-    }
-
-    fn ident_expr(name: &str) -> ast::Expr {
-        ast::Expr::Ident(ast::Ident {
-            span: DUMMY_SP,
-            ctxt: Default::default(),
-            sym: name.into(),
-            optional: false,
-        })
-    }
-
-    fn member_expr(obj: ast::Expr, prop: &str) -> ast::Expr {
-        ast::Expr::Member(ast::MemberExpr {
-            span: DUMMY_SP,
-            obj: Box::new(obj),
-            prop: ast::MemberProp::Ident(ast::IdentName {
-                span: DUMMY_SP,
-                sym: prop.into(),
-            }),
-        })
-    }
-
-    fn call_expr(callee: ast::Expr, args: Vec<ast::Expr>) -> ast::Expr {
-        ast::Expr::Call(ast::CallExpr {
-            span: DUMMY_SP,
-            ctxt: Default::default(),
-            callee: ast::Callee::Expr(Box::new(callee)),
-            args: args
-                .into_iter()
-                .map(|e| ast::ExprOrSpread {
-                    spread: None,
-                    expr: Box::new(e),
-                })
-                .collect(),
-            type_args: None,
-        })
-    }
-
-    fn num_lit(n: f64) -> ast::Expr {
-        ast::Expr::Lit(ast::Lit::Num(ast::Number {
-            span: DUMMY_SP,
-            value: n,
-            raw: None,
-        }))
-    }
-
-    #[test]
-    fn format_expr_recognizes_string_cast() {
-        // `String(entry.totalClicks)`
-        let expr = call_expr(
-            ident_expr("String"),
-            vec![member_expr(ident_expr("entry"), "totalClicks")],
-        );
-        let fmt = try_parse_widget_format_expr(&expr).expect("expected Formatted match");
-        assert!(matches!(fmt.call, WidgetFormatCall::StringCast));
-        assert!(matches!(
-            fmt.arg,
-            WidgetFormatArg::Field(ref f) if f == "totalClicks"
-        ));
-    }
-
-    #[test]
-    fn format_expr_recognizes_math_round_floor_ceil() {
-        for (method, expected) in [
-            ("round", WidgetFormatCall::Round),
-            ("floor", WidgetFormatCall::Floor),
-            ("ceil", WidgetFormatCall::Ceil),
-        ] {
-            let callee = member_expr(ident_expr("Math"), method);
-            let expr = call_expr(
-                callee,
-                vec![member_expr(ident_expr("entry"), "totalClicks")],
-            );
-            let fmt =
-                try_parse_widget_format_expr(&expr).expect("expected Math.* to match whitelist");
-            assert!(
-                std::mem::discriminant(&fmt.call) == std::mem::discriminant(&expected),
-                "method `{}` produced {:?}, want {:?}",
-                method,
-                fmt.call,
-                expected
-            );
-        }
-    }
-
-    #[test]
-    fn format_expr_recognizes_to_fixed_with_digits_literal() {
-        // `entry.totalClicks.toFixed(2)`
-        let target = member_expr(ident_expr("entry"), "totalClicks");
-        let callee = member_expr(target, "toFixed");
-        let expr = call_expr(callee, vec![num_lit(2.0)]);
-        let fmt = try_parse_widget_format_expr(&expr).expect("expected toFixed match");
-        assert!(matches!(fmt.call, WidgetFormatCall::ToFixed { digits: 2 }));
-        assert!(matches!(
-            fmt.arg,
-            WidgetFormatArg::Field(ref f) if f == "totalClicks"
-        ));
-    }
-
-    #[test]
-    fn format_expr_rejects_unknown_call_shape() {
-        // `fmt(entry.totalClicks)` — user-defined function, not in
-        // the whitelist; must return None (caller falls back to an
-        // empty literal rather than producing broken output).
-        let expr = call_expr(
-            ident_expr("fmt"),
-            vec![member_expr(ident_expr("entry"), "totalClicks")],
-        );
-        assert!(try_parse_widget_format_expr(&expr).is_none());
-    }
-
-    #[test]
-    fn parse_text_content_call_path_collapses_for_unknown() {
-        // Same as above but exercising the public entrypoint — verifies
-        // that the call falls through to Literal("") rather than
-        // silently dropping into Field("").
-        let expr = call_expr(
-            ident_expr("fmt"),
-            vec![member_expr(ident_expr("entry"), "totalClicks")],
-        );
-        match parse_text_content(&expr) {
-            WidgetTextContent::Literal(ref s) if s.is_empty() => {}
-            other => panic!("expected Literal(\"\"), got {:?}", other),
-        }
-    }
-
-    #[test]
-    fn parse_text_content_call_path_recognized_call_becomes_formatted() {
-        // `Math.round(entry.totalClicks)` — the user's actual bug
-        // shape. Must produce Formatted, not the catch-all empty
-        // literal that lost the call body.
-        let expr = call_expr(
-            member_expr(ident_expr("Math"), "round"),
-            vec![member_expr(ident_expr("entry"), "totalClicks")],
-        );
-        match parse_text_content(&expr) {
-            WidgetTextContent::Formatted(fmt) => {
-                assert!(matches!(fmt.call, WidgetFormatCall::Round));
-                assert!(matches!(
-                    fmt.arg,
-                    WidgetFormatArg::Field(ref f) if f == "totalClicks"
-                ));
-            }
-            other => panic!("expected Formatted, got {:?}", other),
-        }
-    }
-
-    /// Parse `source`, extract the body statements of the first function
-    /// declaration, and run the reload-policy scanner over them.
-    fn scan_first_fn_body(source: &str) -> (Vec<u32>, bool) {
-        let module = perry_parser::parse_typescript(source, "widget_reload_test.ts")
-            .expect("test source parses");
-        for item in &module.body {
-            if let ast::ModuleItem::Stmt(ast::Stmt::Decl(ast::Decl::Fn(fn_decl))) = item {
-                let body = fn_decl.function.body.as_ref().expect("fn has body");
-                let mut found = Vec::new();
-                let mut unparsed = false;
-                scan_provider_stmts_for_reload_policy(&body.stmts, &mut found, &mut unparsed);
-                return (found, unparsed);
-            }
-        }
-        panic!("no function declaration in test source");
-    }
-
-    #[test]
-    fn reload_policy_literal_minutes_becomes_seconds() {
-        let (found, unparsed) = scan_first_fn_body(
-            r#"
-            async function provider() {
-                return {
-                    entries: [{ temperature: 20 }],
-                    reloadPolicy: { after: { minutes: 30 } },
-                };
-            }
-            "#,
-        );
-        assert_eq!(found, vec![1800]);
-        assert!(!unparsed);
-    }
-
-    #[test]
-    fn reload_policy_found_inside_if_and_try() {
-        // The docs' error-handling shape: a short retry interval on the
-        // failure path, a longer one on the happy path.
-        let (found, unparsed) = scan_first_fn_body(
-            r#"
-            async function provider() {
-                try {
-                    const res = await fetch("https://api.example.com/weather");
-                    if (!res.ok) {
-                        return {
-                            entries: [{ temperature: 0 }],
-                            reloadPolicy: { after: { minutes: 5 } },
-                        };
-                    }
-                    return {
-                        entries: [{ temperature: 21 }],
-                        reloadPolicy: { after: { minutes: 15 } },
-                    };
-                } catch (e) {
-                    return {
-                        entries: [{ temperature: 0 }],
-                        reloadPolicy: { after: { minutes: 1 } },
-                    };
-                }
-            }
-            "#,
-        );
-        assert_eq!(found, vec![300, 900, 60]);
-        assert!(!unparsed);
-    }
-
-    #[test]
-    fn reload_policy_non_literal_flags_unparsed() {
-        let (found, unparsed) = scan_first_fn_body(
-            r#"
-            async function provider() {
-                const policy = { after: { minutes: 10 } };
-                return { entries: [], reloadPolicy: policy };
-            }
-            "#,
-        );
-        assert!(found.is_empty());
-        assert!(unparsed);
-    }
-
-    #[test]
-    fn reload_policy_missing_is_neither_found_nor_unparsed() {
-        let (found, unparsed) = scan_first_fn_body(
-            r#"
-            async function provider() {
-                return { entries: [{ temperature: 20 }] };
-            }
-            "#,
-        );
-        assert!(found.is_empty());
-        assert!(!unparsed);
-    }
-
-    #[test]
-    fn reload_policy_arrow_expression_body_scanned() {
-        // `provider: async () => ({ entries: [], reloadPolicy: ... })` —
-        // exercise the expression-body scanner directly.
-        let module = perry_parser::parse_typescript(
-            r#"const p = async () => ({ entries: [], reloadPolicy: { after: { minutes: 45 } } });"#,
-            "widget_reload_arrow_test.ts",
-        )
-        .expect("test source parses");
-        let mut found = Vec::new();
-        let mut unparsed = false;
-        for item in &module.body {
-            if let ast::ModuleItem::Stmt(ast::Stmt::Decl(ast::Decl::Var(var))) = item {
-                let init = var.decls[0].init.as_ref().expect("var has init");
-                let ast::Expr::Arrow(arrow) = init.as_ref() else {
-                    panic!("expected arrow init");
-                };
-                let ast::BlockStmtOrExpr::Expr(expr) = arrow.body.as_ref() else {
-                    panic!("expected expression body");
-                };
-                scan_provider_return_expr_for_reload_policy(expr, &mut found, &mut unparsed);
-            }
-        }
-        assert_eq!(found, vec![2700]);
-        assert!(!unparsed);
-    }
-
-    #[test]
-    fn reload_policy_seconds_parser_edge_cases() {
-        // Fractional minutes round to the nearest second.
-        let half_minute = object_lit(vec![key_value(
-            "after",
-            object_lit(vec![key_value("minutes", num_lit(0.5))]),
-        )]);
-        assert_eq!(parse_reload_policy_seconds(&half_minute), Some(30));
-
-        // Zero and negative intervals are rejected.
-        let zero = object_lit(vec![key_value(
-            "after",
-            object_lit(vec![key_value("minutes", num_lit(0.0))]),
-        )]);
-        assert_eq!(parse_reload_policy_seconds(&zero), None);
-        let negative = object_lit(vec![key_value(
-            "after",
-            object_lit(vec![key_value("minutes", num_lit(-5.0))]),
-        )]);
-        assert_eq!(parse_reload_policy_seconds(&negative), None);
-
-        // Missing `minutes`, wrong key, or non-object shapes are rejected.
-        let empty_after = object_lit(vec![key_value("after", object_lit(vec![]))]);
-        assert_eq!(parse_reload_policy_seconds(&empty_after), None);
-        let wrong_key = object_lit(vec![key_value(
-            "every",
-            object_lit(vec![key_value("minutes", num_lit(10.0))]),
-        )]);
-        assert_eq!(parse_reload_policy_seconds(&wrong_key), None);
-        assert_eq!(parse_reload_policy_seconds(&str_lit("hourly")), None);
-    }
-
-    #[test]
-    fn entry_field_value_spec_array_with_string_suffix_spec() {
-        // `'string[]'` and `['string']` both resolve to `Array<String>`.
-        let by_suffix = parse_entry_field_value_spec(&str_lit("string[]"));
-        let by_literal = parse_entry_field_value_spec(&array_lit(vec![str_lit("string")]));
-        assert!(matches!(
-            by_suffix,
-            WidgetFieldType::Array(inner) if matches!(*inner, WidgetFieldType::String)
-        ));
-        assert!(matches!(
-            by_literal,
-            WidgetFieldType::Array(inner) if matches!(*inner, WidgetFieldType::String)
-        ));
-    }
-}
+mod tests;
